Changes to how public flags work
